+++ web/experimental/embed-new.html
@@ -14,7 +14,7 @@ BSD-style license that can be found in the LICENSE file. -->
 
     <title>DartPad</title>
 
-<script defer src="scripts/new_embed.dart.js"></script>
+<script defer src="new_embed.dart.js"></script>
 </head>
 
 <body>
